app/requirements.txt
bash==0.6
bcrypt==4.0.1
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dominate==2.8.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyodbc==4.0.39
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
urllib3==2.0.4
uWSGI==2.0.21
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
